refactor: move auth_textfield to widgets folder
