+++ lib/screens/common_widgets/auth/basic_auth_fields.dart
@@ -1,4 +1,4 @@
-import 'package:apidash/screens/common_widgets/auth_textfield.dart';
+import 'package:apidash/widgets/auth_textfield.dart';
 import 'package:flutter/material.dart';
 import 'package:apidash_core/apidash_core.dart';
